+++ src/main.py
@@ -1,4 +1,4 @@
-# Turbulence
+# Turbulence - Chunkless
 # Made With LOVE by kreglar XOXOXOXO!!
 # Version 1.0.2