+++ src/utils/project.py
@@ -11,16 +11,14 @@ class ProjectData:
     # define sets and maps
     palettes: list[data.Palette]
     tileset: data.Tileset
-    chunkset: data.Chunkset
     tilemap: data.Tilemap
 
 def NewProjectFile() -> ProjectData:
     # define defaults
     palettes = [data.Palette() for i in range(4)]
     tileset = data.Tileset(1600) # 1600 unique tiles
-    chunkset = data.Chunkset(400, 4) # 400 unique chunks, 4x4 tiles per chunk
-    tilemap = data.Tilemap((64, 32)) # 64x32 chunk tilemap
+    tilemap = data.Tilemap((256, 128)) # 256x128 tile tilemap
 
     # create project dataclass
-    projectFile = ProjectData(palettes, tileset, chunkset, tilemap)
+    projectFile = ProjectData(palettes, tileset, tilemap)
     return projectFile