+++ CHANGELOG.md
@@ -4,17 +4,17 @@ All notable changes to this project will be recorded here.
 
 ## [1.0.2] - 7/17/25 - Bug Fixes/Demo Files
 
-- Fixes importing chunksets and tilemaps
+- Fixes importing tilemaps
 - added demo files for testing
 
 ## [1.0.1] - 7/15/25 - Bug Fixes/Documentation
 
-- Fixes exporting chunksets and tilemaps.
+- Fixes exporting tilemaps.
 - Documentation and ReadMe updates
 
 ## [1.0.0] - 7/1/25 - Release
 
-- palettes, tilesets, chunksets, and tilemaps
+- palettes, tilesets, and tilemaps
     - create
     - edit
     - import